chore(Languages/CCS): Mathlib linting, more readable proof of Act.Co decidability (#113)
This file did not depend on Mathlib, which given the use of `weak` in
`lakefile.toml` meant that this was silently not linted. There are three
other files for which this is true:
```
Cslib.Foundations.Syntax.HasAlphaEquiv
Cslib.Foundations.Syntax.HasSubstitution
Cslib.Foundations.Syntax.HasWellFormed
```
but they only define notation typeclasses and have no lint problems.

While here, I also took the opportunity to make a decidability proof
easier to read.