feat: Support extensional equality in FinFun (#118)
This PR changes `FinFun` to support extensional equality, making it
essentially equivalent to `Finsupp` in Mathlib, with the important
difference that definitions and theories about `FinFun` are computable.
It also makes FinFun adopt the same notation as Finsupp, so to
facilitate a future replacement with the latter in the future if it ever
changes in Mathlib to have computable defs/theorems (this is relegated
to the future since FinFun is used downstream in at least one big
project I know of).

---------

Co-authored-by: Chris Henson <46805207+chenson2018@users.noreply.github.com>