feat: add Cslib/Computability/Languages/OmegaLanguage.lean (#92)
This is the first version of OmegaLanguage, which is an analogue of
Mathlib.Computability.Language, but for infinite sequences. We now have
a more satisfactory definition of omegaPow, which enables more algebraic
and much shorter proofs of the harder results (see the "main theorems"
section of the main comment in OmegaLanguage.lean) than in:

https://github.com/ctchou/AutomataTheory/blob/main/AutomataTheory/Languages/Basic.lean
The results proved so far are sufficient for starting the development of
the theory of omega-automata. The main TODOs are results about omegaPow
and map, which are not needed for now.

As part of this work, we added several new definitions and theorems
about Language and OmegaSequence.flatten. We also took this opportunity
to change some infelicitous theorem names in Flatten.lean and
Segment.lean and to use `namespace Nat` to remove the prefix `Nat.` from
many places in Segment.lean.

---------

Co-authored-by: Chris Henson <chrishenson.net@gmail.com>