feat: new design for automata theory, results on regular languages, and deterministic labelled transition systems (#144)
This PR merges the developments that @ctchou and I carried out in the
`automata` branch, #141, and #142, as discussed on Zulip ([#CSLib >
Question Lean structure and
extends](https://leanprover.zulipchat.com/#narrow/channel/513188-CSLib/topic/Question.20Lean.20structure.20and.20extends/with/554287177)).

The design based on `Acceptor` and structure extension works pretty
well. Some definitions could be removed because of the `Acceptor`
typeclass, and some proofs got easier thanks to the combination of
structure extension, record updates, and grind.

Co-authored with @ctchou.

---------

Co-authored-by: Ching-Tsun Chou <chingtsun.chou@gmail.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Chris Henson <46805207+chenson2018@users.noreply.github.com>